# perf: ⚡️ 30% about normalize