main branch:
```bash
Gnuplot not found, using plotters backend
join_impl               time:   [38.802 us 39.369 us 40.077 us]                       
                        change: [-21.120% -20.056% -18.612%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  2 (2.00%) high mild
```
current branch:
```bash
Gnuplot not found, using plotters backend
join_impl               time:   [26.758 us 27.181 us 27.645 us]                       
                        change: [-32.062% -30.856% -29.763%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
```